refactor: split Dock unread projection source
